pyproject.toml
[tool.poetry]
name = "fastapi-gdelt-industrial"
version = "0.1.0"
description = "FastAPI haute performance POC -> industriel avec ingestion GDELT big data"
authors = ["You <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic-settings = "^2.5.2"
httpx = "^0.27.2"
tenacity = "^8.5.0"
python-json-logger = "^2.0.7"
prometheus-client = "^0.20.0"
sqlalchemy = "^2.0.34"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
redis = "^5.0.8"
arq = "^0.26.1"
boto3 = "^1.35.10"
pyarrow = "^17.0.0"
duckdb = "^1.0.0"
pandas = "^2.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.8"

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
